requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.5.0
oauth2client==4.1.3
oauthlib==3.3.0
Panda3D==1.10.15
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-telegram-bot==22.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1
